Refactor: shareRoom -> selectedArea 로 파일명 변경
shareRoom -> selectedArea로 파일명 변경
