+++ src/state/selectedArea.ts
@@ -1,0 +1,6 @@
+import { atom } from "recoil";
+
+export const selectedArea = atom({
+    key: "selectedArea",
+    default: "noSelected"
+})
